% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDS_Pred.R
\name{autoMLpredict}
\alias{autoMLpredict}
\title{Use build models to predict CDS expression value}
\usage{
autoMLpredict(filenm)
}
\arguments{
\item{infile}{DNA sequence file}
}
\value{
None
}
\description{
Use build models to predict CDS expression value
}
