% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDS_Pred.R
\name{CodonCount}
\alias{CodonCount}
\title{Create a file with Codon count numbers}
\usage{
CodonCount(filenm)
}
\arguments{
\item{infile}{DNA sequence file}
}
\value{
None (new file with "*_codonSEQ.txt" will be created)
}
\description{
Create a file with Codon count numbers
}
