% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDS_Pred.R
\name{summaryPredict}
\alias{summaryPredict}
\title{Get the summary data file}
\usage{
summaryPredict(filenm)
}
\arguments{
\item{infile}{DNA sequence file}
}
\value{
None
}
\description{
Get the summary data file
}
